UI: renaming BanditConfig.jsx
