+++ src/components/BanditConfig.jsx
